Mover timer functions in separate C file
