+++ regex.c
@@ -11,7 +11,7 @@
 #include "FTL.h"
 #include "regex_r.h"
 #include "gravity.h"
-#include "daemon.h"
+#include "timers.h"
 #include "memory.h"
 #include "log.h"
 #include <regex.h>
