+++ request.c
@@ -11,7 +11,7 @@
 #include "FTL.h"
 #include "api.h"
 #include "shmem.h"
-#include "daemon.h"
+#include "timers.h"
 #include "request.h"
 #include "socket.h"
 #include "resolve.h"
